+++ go-client-sdk/pkg/models/shared/codatpublicapimodelscompanyrenamecompanymodel.go
@@ -1,5 +1,0 @@
-package shared
-
-type CodatPublicAPIModelsCompanyRenameCompanyModel struct {
-	Name *string `json:"name,omitempty"`
-}