+++ go-client-sdk/pkg/models/shared/codatpublicapimodelscompanyupdatecompanymodel.go
@@ -1,0 +1,6 @@
+package shared
+
+type CodatPublicAPIModelsCompanyUpdateCompanyModel struct {
+	Description *string `json:"description,omitempty"`
+	Name        *string `json:"name,omitempty"`
+}