+++ python-client-sdk/src/sdk/assess.py
@@ -218,6 +218,31 @@ class Assess:
         return res
 
     
+    def get_data_companies_company_id_assess_excel_download(self, request: operations.GetDataCompaniesCompanyIDAssessExcelDownloadRequest) -> operations.GetDataCompaniesCompanyIDAssessExcelDownloadResponse:
+        r"""Download the Excel report to a local drive.
+        """
+        
+        base_url = self._server_url
+        
+        url = utils.generate_url(base_url, "/data/companies/{companyId}/assess/excel/download", request.path_params)
+        
+        query_params = utils.get_query_params(request.query_params)
+        
+        client = utils.configure_security_client(self._client, request.security)
+        
+        r = client.request("GET", url, params=query_params)
+        content_type = r.headers.get("Content-Type")
+
+        res = operations.GetDataCompaniesCompanyIDAssessExcelDownloadResponse(status_code=r.status_code, content_type=content_type)
+        
+        if r.status_code == 200:
+            if utils.match_content_type(content_type, "application/json"):
+                out = utils.unmarshal_json(r.text, Optional[bytes])
+                res.get_data_companies_company_id_assess_excel_download_200_application_json_binary_string = out
+
+        return res
+
+    
     def get_data_companies_company_id_connections_connection_id_assess_accounting_metrics_marketing(self, request: operations.GetDataCompaniesCompanyIDConnectionsConnectionIDAssessAccountingMetricsMarketingRequest) -> operations.GetDataCompaniesCompanyIDConnectionsConnectionIDAssessAccountingMetricsMarketingResponse:
         r"""Gets the marketing metrics from an accounting source for a given company, over one or more periods of time.
         """