+++ python-client-sdk/src/sdk/sdk.py
@@ -125,8 +125,8 @@ class SDK:
     _security: shared.Security
     _server_url: str = SERVERS[0]
     _language: str = "python"
-    _sdk_version: str = "1.1.4"
-    _gen_version: str = "0.21.2"
+    _sdk_version: str = "1.2.0"
+    _gen_version: str = "0.22.0"
 
     def __init__(self) -> None:
         self._client = requests.Session()