+++ typescript-client-sdk/src/sdk/models/shared/codatpublicapimodelscompanyrenamecompanymodel.ts
@@ -1,8 +1,0 @@
-import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
-
-
-
-export class CodatPublicApiModelsCompanyRenameCompanyModel extends SpeakeasyBase {
-  @SpeakeasyMetadata({ data: "json, name=name" })
-  name?: string;
-}