+++ typescript-client-sdk/src/sdk/models/shared/codatpublicapimodelscompanyupdatecompanymodel.ts
@@ -2,13 +2,10 @@ import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
 
 
 
-export class CodatPublicApiModelsCompanyAddCompanyModel extends SpeakeasyBase {
+export class CodatPublicApiModelsCompanyUpdateCompanyModel extends SpeakeasyBase {
   @SpeakeasyMetadata({ data: "json, name=description" })
   description?: string;
 
   @SpeakeasyMetadata({ data: "json, name=name" })
-  name: string;
-
-  @SpeakeasyMetadata({ data: "json, name=platformType" })
-  platformType?: string;
+  name?: string;
 }