+++ RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [NPM v1.1.4] https://www.npmjs.com/package/@codatio/codat-ts/v/1.1.4 - typescript-client-sdk
 - [PyPI v1.1.4] https://pypi.org/project/codatapi/1.1.4 - python-client-sdk
 - [Go v1.1.4] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.1.4 - go-client-sdk
+
+## Version 1.2.0
+### Changes
+Based on:
+- OpenAPI Doc v1 https://api.codat.io/swagger/v1/swagger.json
+- Speakeasy CLI 0.22.0 https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v1.2.0] https://www.npmjs.com/package/@codatio/codat-ts/v/1.2.0 - typescript-client-sdk
+- [PyPI v1.2.0] https://pypi.org/project/codatapi/1.2.0 - python-client-sdk
+- [Go v1.2.0] https://github.com/speakeasy-sdks/codat-sdks/releases/tag/v1.2.0 - go-client-sdk